+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+## 0.0.3
+
+- Support for multiple commands in one line
+- Support for commands on multiple lines
+- Adding quiet mode
+- Adding parallel mode
+
 ## 0.0.2
 
-- Fix for nested `;`
+- Fix for nested commands
 
 ## 0.0.1